+++ examples/misc.rs
@@ -1,42 +1,26 @@
-use fred::{
-  prelude::*,
-  types::{BackpressureConfig, BackpressurePolicy, PerformanceConfig},
-};
+use fred::prelude::*;
+use std::time::Duration;
 
 #[tokio::main]
 async fn main() -> Result<(), RedisError> {
-  // full configuration for performance tuning options
-  let perf = PerformanceConfig {
-    // whether or not to automatically pipeline commands across tasks
-    auto_pipeline:                                             true,
-    // the max number of frames to feed into a socket before flushing it
-    max_feed_count:                                            1000,
-    // a default timeout to apply to all commands (0 means no timeout)
-    default_command_timeout_ms:                                0,
-    // the amount of time to wait before rebuilding the client's cached cluster state after a MOVED error.
-    cluster_cache_update_delay_ms:                             10,
-    // the maximum number of times to retry commands
-    max_command_attempts:                                      3,
-    // backpressure config options
-    backpressure:                                              BackpressureConfig {
-      // whether to disable automatic backpressure features
-      disable_auto_backpressure: false,
-      // the max number of in-flight commands before applying backpressure or returning backpressure errors
-      max_in_flight_commands:    5000,
-      // the policy to apply when the max in-flight commands count is reached
-      policy:                    BackpressurePolicy::Drain,
-    },
-    // the amount of time a command can wait in memory without a response before the connection is considered
-    // unresponsive
-    #[cfg(feature = "check-unresponsive")]
-    network_timeout_ms:                                        60_000,
-  };
-  let config = RedisConfig {
-    server: ServerConfig::default_clustered(),
-    ..RedisConfig::default()
-  };
-
-  let client = RedisClient::new(config, Some(perf), None);
+  let client = Builder::default_centralized()
+    .with_performance_config(|config| {
+      config.max_feed_count = 1000;
+      config.auto_pipeline = true;
+    })
+    .with_connection_config(|config| {
+      config.tcp = TcpConfig {
+        nodelay: true,
+        ..Default::default()
+      };
+      config.max_command_attempts = 5;
+      config.max_redirections = 5;
+      config.internal_command_timeout_ms = 2_000;
+      config.connection_timeout_ms = 10_000;
+    })
+    // use exponential backoff, starting at 100 ms and doubling on each failed attempt up to 30 sec
+    .set_policy(ReconnectPolicy::new_exponential(0, 100, 30_000, 2))
+    .build()?;
   let _ = client.connect();
   let _ = client.wait_for_connect().await?;
 
@@ -58,9 +42,23 @@ async fn main() -> Result<(), RedisError> {
 
   // update performance config options
   let mut perf_config = client.perf_config();
-  perf_config.max_command_attempts = 100;
   perf_config.max_feed_count = 1000;
   client.update_perf_config(perf_config);
+
+  // overwrite configuration options on individual commands
+  let options = Options {
+    max_attempts: Some(5),
+    max_redirections: Some(5),
+    timeout: Some(Duration::from_secs(10)),
+    ..Default::default()
+  };
+  let _: Option<String> = client.with_options(&options).get("foo").await?;
+
+  // apply custom options to a pipeline
+  let pipeline = client.pipeline().with_options(&options);
+  let _: () = pipeline.get("foo").await?;
+  let _: () = pipeline.get("bar").await?;
+  let (_, _): (Option<i64>, Option<i64>) = pipeline.all().await?;
 
   // interact with specific cluster nodes
   if client.is_clustered() {
@@ -69,9 +67,13 @@ async fn main() -> Result<(), RedisError> {
     for server in connections.into_iter() {
       let info: String = client.with_cluster_node(&server).client_info().await?;
       println!("Client info for {}: {}", server, info);
+
+      // or combine client wrappers
+      let _: () = client.with_cluster_node(&server).with_options(&options).ping().await?;
     }
   }
 
   let _ = client.quit().await?;
+  let _ = events_task.await;
   Ok(())
 }